Enable more GCC warning flags

+++ src/neo/fft/dft_test.cpp
@@ -61,8 +61,8 @@ auto test_dft_plan()
 
             auto const n     = static_cast<double>(size);
             auto const scale = neo::bit_log2(size) * 4;
-            REQUIRE(x[i].real() == Catch::Approx(sig[i].real() * n).scale(scale));
-            REQUIRE(x[i].imag() == Catch::Approx(sig[i].imag() * n).scale(scale));
+            REQUIRE(x[i].real() == Catch::Approx(double(sig[i].real()) * n).scale(scale));
+            REQUIRE(x[i].imag() == Catch::Approx(double(sig[i].imag()) * n).scale(scale));
         }
     }
 }
